LayoutTests/fast/webgpu/nocrash/fuzz-145584310-expected.txt

CONSOLE MESSAGE: promises created
CONSOLE MESSAGE: the end
CONSOLE MESSAGE: fuzz-145584310.html
